Refactor regime to its own module
