+++ tradeexecutor/visual/bullbear.py
@@ -3,70 +3,12 @@
 - Visualise discreet bull/bear market filter by showing coloured regions over price chart
 
 """
-import enum
-from dataclasses import dataclass
-from typing import Iterable
 
 import pandas as pd
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 
-
-class Regime(enum.Enum):
-    """Possible market regimes.
-
-    The content of signal for :py:func:`visualise_market_regime_filter`
-    """
-    bull = 1
-    bear = -1
-    crab = 0
-
-
-@dataclass(frozen=True, slots=True)
-class RegimeRegion:
-    """One regime colouring region for the charts."""
-    start: pd.Timestamp
-    end: pd.Timestamp
-    regime: Regime
-
-
-def get_regime_signal_regions(signal: pd.Series) -> Iterable[RegimeRegion]:
-    """Get regions of the regime signal.
-
-    Split the signal to continous blocks for coloring.
-
-    :return:
-        Iterable of market regimes for colouring
-    """
-
-    # https://stackoverflow.com/a/69222703/315168
-    edges = signal.diff(periods=1)
-
-    # edge_mask = edges.loc[edges != 0]
-
-    current_signal = Regime(signal.iloc[0])
-    current_start = edges.index[0]
-
-    regime_change_timestamps = edges.index[edges != 0]
-
-    if len(regime_change_timestamps) > 0:
-        # Skip the start region
-        for idx in regime_change_timestamps[1:]:
-            yield RegimeRegion(
-                current_start,
-                idx,
-                current_signal
-            )
-            current_start = idx
-            current_signal = Regime(signal[idx])
-
-    # The closing region
-    yield RegimeRegion(
-        current_start,
-        signal.index[-1],
-        current_signal,
-    )
-
+from tradeexecutor.analysis.regime import Regime, get_regime_signal_regions
 
 #: Different regime colours
 DEFAULT_COLOUR_MAP = {
